Fix issue 9 (#10)
* Fix #9
Now the script asks the username and the password for every combination of machine:environment the first time it is provisioned. It also assigns a pseudo-random UID to the user to avoid conflicts caused by multiple users (unlikely scenario).

* Fix #9
Right now the 'build' script asks for each host machine the password to assign to the 'discos' user, which it's been decided to leave static, along with its UID (3060). The script asks for the password only if the user does not already exist in the target machine (which means only the first time that machine is provisioned).

* Fix #9 after review https://github.com/discos/deployment/pull/10#pullrequestreview-67858411.