Fix issue 103 (#108)
* Fix #103, fixed lustre installation and procedure.
Added a script called `lustre-startup` that handles the lustre configuration and mounting process. It gets added to startup scripts so that the lustre disk gets mounted automatically. The script also checks if the lustre disk is already mounted. In that case it does nothing.

* Fix #103, improved lustre-startup script.

* Fix #103, lustre installation and mounting process should now be completely idempotent.
Some tweaks were necessary to completely fix the issue.