Fix #105, #106, #112, #113, #114, #115, #116, #117. (#118)
This commit is a major rework of the deployment process. Now it performs some different actions for `manager` and `console` machines. Down below you can find an exhaustive changelog. With these modifications, some minor tweaks were added not present in the changelog.

Fix #105, the directory tree under `archive` gets created in the same way as before, but the lustre remote disk gets mounted under `archive/data`, so that data created when observing gets stored properly in the storage machine.

Fix #106, VNC server gets now installed in all machines. It does not get configured yet because we are missing the user adding mechanism (VNC will be enabled only for users/observers and not for discos).

Fix #112, a log rotation mechanism is now configured in manager machine in order to avoid ACS using a single file that becomes huge and useless after some time.

Fix #113, moved ACS_TMP environment variable from `alma/...` subdirectory to `<user home>/.acs/tmp/<machine name>`.

Fix #114, moved `discos` user home directory to `/discos/discos/`. The `/home/` directory is now empty so observers and users will be the only one appearing in that directory in the future. It it suggested to create the `/discos` directory onto a different physical disk, which should be mounted exactly into the `/discos` path.

Fix #115, an ansible role to deploy console machine(s) is been added to the deployment process. Now it performs a minimal configuration of VNC by applying a patch to GNOME to avoid an annoying popup when logging in via VNC.

Fix #116, the lustre yum packages were installed on the fly by pointing yum to their remote url. Whenever these packages will not be available online anymore, this behavior could be problematic. Now these packages gets downloaded to the local repository before installation, so we can store them in a dedicated repository machine.

Fix #117, prospector was failing to install. Pip was trying to install the latest version, which I believe it has some sort of bug or it gets along Python 3+, which we do not use (we use 2.7 still). I picked the last working version (0.12.11) and its installation now gets completed properly.