Fix #124, grant access to the introot to observer
A major rework of the whole process has been made in order to create the
introot outside the `discos` user home directory. Access to the introot
is now granted with read and execute permissions to all acs group
members. This will enable a user `observer`, which should be created in
console machine, to launch the discos console.