Fix #126, reworked installation procedure.
Now the installation procedure does not rely on python virtual
environments. Instead, it sources the scripts present in the `scripts`
directory by adding the following line in the user's `~/.bashrc` file:
`source $HOME/<path-to-the-repository-directory>/scripts/source.sh`.
I also added a new script called `discos-vnc`, which establishes the ssh
secure tunnel to the specified virtual machine in order to correctly
open the desired VNC session (for `discos` or `observer` user, only in
console machine).