Fix #119, added VNC configuration. (#128)
I decided to allow VNC sessions only when a secure ssh tunnel is
established. I also disabled the VNC password for both users (`discos`
and `observer`) since using the same password for both login and VNC is
a security flaw (VNC stores the password with a VERY weak encryption).
By forcing the remote users to type the `observer` login password to
establish the ssh tunnel, the VNC password is not required anymore. I
also defined a 'starting port' for VNC displays to everride the default
one (which is 5900+display number), since the default one collides with
some of the systems installed in the SRT station (and maybe in other
stations too). In order to prevent `observer` users to hijack the
`discos` user VNC session, I also configured a rule to allow `observer`
user to establish a ssh secure tunnel only for its VNC display port,
denying every other connection. I don't really know if this could become
an issue in the future, this must be discussed.