Fix #135, #136, improved `discos-deploy` script. (#137)
Now the script accepts an optional argument: `-n` or `--no-vagrant`.
It disabled the calls to Vagrant for the current execution. This allows
an user to deploy the development environment onto machines (virtual or
physical) without the need of bringing them running with Vagrant.
The script also checks if all the target machines are reachable before
calling Ansible, and exits with and error if they are not reachable.