Fix #134, added a storage machine to the process. (#140)
Now the process is able to install and configure a storage server, which
operates both as a lustre server and a NIS server. I apologize for the
large number of files edited included in this commit, but I tried to
setup a truly common strategy for every machine. The old Ansible common
role has now been splitted into a truly common role (with tasks useful
for both Centos 6.8 and 7.2 distros), and a ACS role (that handles the
whole ACS + DISCOS setup as the old common role).

Please take note that the lustre server procedure handles the lustre
disk configuration (it formats the specified disk), and that it combines
the lustre MGS and MDT into a single node. This can be further customized
in the future.