Fix #185, fix #186, fix #187, see changelog below. (#189)
Fix #185, reviewed the logrotation mechanism. Now the `acs.xml` file
gets immediately copied with the date on the name, it gets compressed
and then it gets copied to a local filesystem directory (the original
resides in the lustre partition).

Fix #186, in order to avoid confusion, the `/discos` directory, the
directory containing all discos-related users and stuff, has been
renamed to `/discos-sw`. Naming a directory with a user name identify a
user home directory, so the only directory named `discos` should be the
`discos` user home. With this in mind, the `/discos-sw/config/discos`
directory has been renamed to `/discos-sw/config/misc`.

Fix #187, the `/data` directory, the one containing the `acstmp`
directory, has been renamed to `/service`. This was due to the fact that
the name `data` should only identify observation's data.