Fix #200, improved the lustre server installation procedure.
The disk initialization is now optional. In order for for the procedure
to perform it, the `initialize_lustre_disk` variable should be set to
true in the `<inventory>/group_vars/storage` file. If this strategy is
adopted, the `lustre_<XXX>_partition` variables can leave out the
`start` and the `end` fields.

The procedure also handles a mixed `mgs` and `mdt` lustre partition. In
order to enable this feature, the `merge_mgs_mdt` variable should be set
to true in the `<inventory>/group_vars/storage` file. Also, variables
`lustre_mgs_partition` and `lustre_mdt_partition` should be replaced by
the `lustre_mgs_mdt_partition`, keeping the same format.