Fix #196, fix #269. See changelog. (#270)
Fix #196, some further splitting was performed in order to be able to
apply only some subset of tasks for different environments. This opens
the possibility to define a custom playbook for each of the Medicina and
Noto environments.

Fix #269, the repository became a proper Python 2.7 package. It will be
installed in your Python environment and configuration files related to
Vagrant and Ansible will now be copied inside `<HOME>/.deployment`. Also,
the Vagrant environment points to this last mentioned directory, in
order to always retain the same DISCOS virtual machines. The
documentation has been updated according to this latest additions.