Fix #318, updated the whole process in order to use our repository (#319)
* Fix #318, updated the whole process in order to use our repository

All dependencies original URLs have been removed from this repository.
They are now stored in a `sources` text file inside the
discos/dependencies repository (which is private). The process now asks
the user to download part of the dependencies repository with their
github credentials. A token for that repository gets downloaded and
stored the first time the provisioning process is launched. That token
is then used to download single packages from the dependencies repository.
Also, minor bugs were fixed.

* Fix #318, fixed the remote repository link as requested