Fix #322, changed implementation file instead of header file.
A temporary variable is now used instead of changing type to the
original one. The original variable was a `ACS::TimeInterval`
variable, which is probably inherited from the `long long unsigned int`
class and extended. In order to avoid any issue related to some methods
that could be called as a `ACS::TimeInterval` variable but not as a
`long long unsigned int` we chose to fix this issue this way instead of
simply changint the said variable type.