Fix #322, now the ACS patches are dynamically applied
The ACS patches are now applied in a dynamic way which takes into account the ACS version and applies only the patches relative to that version. If we migrate to a new version of ACS that requires no patches, nothing else has to be done, since the procedure will not find any yml file containing the required patches, it will skip to the next task and go on until the procedure is complete. If the new version requires some different patches instead, a new yml file under `/ansible/roles/acs/tasks/patches/` has to be written following this convention for the name :`<acs_tag>.yml` where `acs_tag` is the new `acs_tag` ansible variable that reflect the new ACS version that is being installed.