Fix #322, patched ACS, suppressed warnings regarding ACS::TimeInterval (#323)
* Fix #322, patched ACS, suppressed warnings regarding ACS::TimeInterval

In the `enumpropAlarm.h` file under
`/alma/ACS-FEB2017/ACSSW/Sources/enumprop/ws/include/`, the `interval_m`
variable was declared as `ACS::TimeInterval`. In the `enumpropAlarm.i`
file, under the same directory, a sscanf operation is performed and a
`long long unsigned int` gets stored inside the `interval_m` variable
(internally, the `ACS::TimeInterval` class is indeed a `long long
unsigned int`). The sscanf operation, at compile time, continuously yields
a warning message about this, cluttering the building process output,
making it hard to notice other warning or error messages. This fix
simply changes the type of the `interval_m` variable to `long long
unsigned int`.

* Fix #322, changed implementation file instead of header file.

A temporary variable is now used instead of changing type to the
original one. The original variable was a `ACS::TimeInterval`
variable, which is probably inherited from the `long long unsigned int`
class and extended. In order to avoid any issue related to some methods
that could be called as a `ACS::TimeInterval` variable but not as a
`long long unsigned int` we chose to fix this issue this way instead of
simply changint the said variable type.

* Fix #322, now the ACS patches are dynamically applied

The ACS patches are now applied in a dynamic way which takes into account the ACS version and applies only the patches relative to that version. If we migrate to a new version of ACS that requires no patches, nothing else has to be done, since the procedure will not find any yml file containing the required patches, it will skip to the next task and go on until the procedure is complete. If the new version requires some different patches instead, a new yml file under `/ansible/roles/acs/tasks/patches/` has to be written following this convention for the name :`<acs_tag>.yml` where `acs_tag` is the new `acs_tag` ansible variable that reflect the new ACS version that is being installed.